+++ lantern/src/Trace.cpp
@@ -21,8 +21,12 @@ void* _lantern_create_traceable_fun (void *(*r_caller)(void *, void *), void* fn
     std::function<Stack(Stack)> tr_fn = [r_caller, fn](Stack x)
     {
         //auto r_fn = *reinterpret_cast<std::function<void*(void*)>*>(fn);
-        auto tmp = new LanternObject<Stack>(x);
-        void* out = (*r_caller)((void *)tmp, fn);
+        auto tmp = LanternObject<Stack>(x);
+        void* out = (*r_caller)((void *)(&tmp), fn);
+        // if the R function call fails itt will return nullptr by convention.
+        if (out == nullptr) 
+            throw std::runtime_error("Error in the R function execution.");
+
         return reinterpret_cast<LanternObject<Stack>*>(out)->get();
     };
 
@@ -30,12 +34,14 @@ void* _lantern_create_traceable_fun (void *(*r_caller)(void *, void *), void* fn
     LANTERN_FUNCTION_END;
 }
 
-void* _lantern_trace_fn (void* fn, void* inputs, void* compilation_unit, bool strict)
+void* _lantern_trace_fn (void* fn, void* inputs, void* compilation_unit, bool strict, void* module, void* name, bool should_mangle)
 {
     LANTERN_FUNCTION_START;
     std::function<Stack(Stack)> fn_ = reinterpret_cast<LanternObject<std::function<Stack(Stack)>>*>(fn)->get();
     Stack inputs_ = reinterpret_cast<LanternObject<Stack>*>(inputs)->get();
     CompilationUnit* cu = reinterpret_cast<CompilationUnit*>(compilation_unit);
+    auto module_ = reinterpret_cast<torch::jit::script::Module *>(module);
+    auto name_ = reinterpret_cast<LanternObject<std::string>*>(name)->get();
 
     std::function<std::string(const torch::autograd::Variable&)> var_fn = [](const torch::autograd::Variable& x) {
         return "";
@@ -45,10 +51,12 @@ void* _lantern_trace_fn (void* fn, void* inputs, void* compilation_unit, bool st
         inputs_,
         fn_,
         var_fn,
-        strict
+        strict, 
+        false,
+        module_
     );
 
-    auto tr_fn = cu->create_function("name", std::get<0>(traced)->graph, true);
+    auto tr_fn = cu->create_function(name_, std::get<0>(traced)->graph, should_mangle);
     
     return (void*) tr_fn;
     LANTERN_FUNCTION_END;