+++ src/RcppExports.cpp
@@ -32060,6 +32060,18 @@ BEGIN_RCPP
     return R_NilValue;
 END_RCPP
 }
+// cpp_jit_script_module_register_module
+void cpp_jit_script_module_register_module(XPtrTorchScriptModule self, XPtrTorchstring name, XPtrTorchScriptModule module);
+RcppExport SEXP _torch_cpp_jit_script_module_register_module(SEXP selfSEXP, SEXP nameSEXP, SEXP moduleSEXP) {
+BEGIN_RCPP
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< XPtrTorchScriptModule >::type self(selfSEXP);
+    Rcpp::traits::input_parameter< XPtrTorchstring >::type name(nameSEXP);
+    Rcpp::traits::input_parameter< XPtrTorchScriptModule >::type module(moduleSEXP);
+    cpp_jit_script_module_register_module(self, name, module);
+    return R_NilValue;
+END_RCPP
+}
 // cpp_jit_script_module_to
 void cpp_jit_script_module_to(XPtrTorchScriptModule self, XPtrTorchDevice device, bool non_blocking);
 RcppExport SEXP _torch_cpp_jit_script_module_to(SEXP selfSEXP, SEXP deviceSEXP, SEXP non_blockingSEXP) {
@@ -32107,15 +32119,73 @@ BEGIN_RCPP
 END_RCPP
 }
 // cpp_jit_script_method_call
-Rcpp::XPtr<XPtrTorchStack> cpp_jit_script_method_call(XPtrTorchScriptMethod self, Rcpp::XPtr<XPtrTorchStack> inputs);
+XPtrTorchStack cpp_jit_script_method_call(XPtrTorchScriptMethod self, XPtrTorchStack inputs);
 RcppExport SEXP _torch_cpp_jit_script_method_call(SEXP selfSEXP, SEXP inputsSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
     Rcpp::traits::input_parameter< XPtrTorchScriptMethod >::type self(selfSEXP);
-    Rcpp::traits::input_parameter< Rcpp::XPtr<XPtrTorchStack> >::type inputs(inputsSEXP);
+    Rcpp::traits::input_parameter< XPtrTorchStack >::type inputs(inputsSEXP);
     rcpp_result_gen = Rcpp::wrap(cpp_jit_script_method_call(self, inputs));
     return rcpp_result_gen;
+END_RCPP
+}
+// cpp_jit_script_module_new
+XPtrTorchScriptModule cpp_jit_script_module_new(XPtrTorchCompilationUnit cu, XPtrTorchstring name);
+RcppExport SEXP _torch_cpp_jit_script_module_new(SEXP cuSEXP, SEXP nameSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< XPtrTorchCompilationUnit >::type cu(cuSEXP);
+    Rcpp::traits::input_parameter< XPtrTorchstring >::type name(nameSEXP);
+    rcpp_result_gen = Rcpp::wrap(cpp_jit_script_module_new(cu, name));
+    return rcpp_result_gen;
+END_RCPP
+}
+// cpp_jit_script_module_add_constant
+void cpp_jit_script_module_add_constant(XPtrTorchScriptModule self, XPtrTorchstring name, XPtrTorchIValue value);
+RcppExport SEXP _torch_cpp_jit_script_module_add_constant(SEXP selfSEXP, SEXP nameSEXP, SEXP valueSEXP) {
+BEGIN_RCPP
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< XPtrTorchScriptModule >::type self(selfSEXP);
+    Rcpp::traits::input_parameter< XPtrTorchstring >::type name(nameSEXP);
+    Rcpp::traits::input_parameter< XPtrTorchIValue >::type value(valueSEXP);
+    cpp_jit_script_module_add_constant(self, name, value);
+    return R_NilValue;
+END_RCPP
+}
+// cpp_jit_script_module_add_method
+void cpp_jit_script_module_add_method(XPtrTorchScriptModule self, Rcpp::XPtr<XPtrTorch> method);
+RcppExport SEXP _torch_cpp_jit_script_module_add_method(SEXP selfSEXP, SEXP methodSEXP) {
+BEGIN_RCPP
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< XPtrTorchScriptModule >::type self(selfSEXP);
+    Rcpp::traits::input_parameter< Rcpp::XPtr<XPtrTorch> >::type method(methodSEXP);
+    cpp_jit_script_module_add_method(self, method);
+    return R_NilValue;
+END_RCPP
+}
+// cpp_jit_script_module_find_constant
+SEXP cpp_jit_script_module_find_constant(XPtrTorchScriptModule self, XPtrTorchstring name);
+RcppExport SEXP _torch_cpp_jit_script_module_find_constant(SEXP selfSEXP, SEXP nameSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< XPtrTorchScriptModule >::type self(selfSEXP);
+    Rcpp::traits::input_parameter< XPtrTorchstring >::type name(nameSEXP);
+    rcpp_result_gen = Rcpp::wrap(cpp_jit_script_module_find_constant(self, name));
+    return rcpp_result_gen;
+END_RCPP
+}
+// cpp_jit_script_module_save
+void cpp_jit_script_module_save(XPtrTorchScriptModule self, XPtrTorchstring path);
+RcppExport SEXP _torch_cpp_jit_script_module_save(SEXP selfSEXP, SEXP pathSEXP) {
+BEGIN_RCPP
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< XPtrTorchScriptModule >::type self(selfSEXP);
+    Rcpp::traits::input_parameter< XPtrTorchstring >::type path(pathSEXP);
+    cpp_jit_script_module_save(self, path);
+    return R_NilValue;
 END_RCPP
 }
 // test_stack
@@ -32463,16 +32533,20 @@ BEGIN_RCPP
 END_RCPP
 }
 // cpp_trace_function
-Rcpp::XPtr<XPtrTorchFunctionPtr> cpp_trace_function(Rcpp::Function fn, XPtrTorchStack inputs, XPtrTorchCompilationUnit compilation_unit, bool strict);
-RcppExport SEXP _torch_cpp_trace_function(SEXP fnSEXP, SEXP inputsSEXP, SEXP compilation_unitSEXP, SEXP strictSEXP) {
+Rcpp::XPtr<XPtrTorchFunctionPtr> cpp_trace_function(Rcpp::Function fn, XPtrTorchStack inputs, XPtrTorchCompilationUnit compilation_unit, XPtrTorchstring name, bool strict, XPtrTorchScriptModule module, bool should_mangle, bool manage_memory);
+RcppExport SEXP _torch_cpp_trace_function(SEXP fnSEXP, SEXP inputsSEXP, SEXP compilation_unitSEXP, SEXP nameSEXP, SEXP strictSEXP, SEXP moduleSEXP, SEXP should_mangleSEXP, SEXP manage_memorySEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
     Rcpp::traits::input_parameter< Rcpp::Function >::type fn(fnSEXP);
     Rcpp::traits::input_parameter< XPtrTorchStack >::type inputs(inputsSEXP);
     Rcpp::traits::input_parameter< XPtrTorchCompilationUnit >::type compilation_unit(compilation_unitSEXP);
+    Rcpp::traits::input_parameter< XPtrTorchstring >::type name(nameSEXP);
     Rcpp::traits::input_parameter< bool >::type strict(strictSEXP);
-    rcpp_result_gen = Rcpp::wrap(cpp_trace_function(fn, inputs, compilation_unit, strict));
+    Rcpp::traits::input_parameter< XPtrTorchScriptModule >::type module(moduleSEXP);
+    Rcpp::traits::input_parameter< bool >::type should_mangle(should_mangleSEXP);
+    Rcpp::traits::input_parameter< bool >::type manage_memory(manage_memorySEXP);
+    rcpp_result_gen = Rcpp::wrap(cpp_trace_function(fn, inputs, compilation_unit, name, strict, module, should_mangle, manage_memory));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -35110,11 +35184,17 @@ static const R_CallMethodDef CallEntries[] = {
     {"_torch_cpp_jit_script_module_is_optimized", (DL_FUNC) &_torch_cpp_jit_script_module_is_optimized, 1},
     {"_torch_cpp_jit_script_module_register_parameter", (DL_FUNC) &_torch_cpp_jit_script_module_register_parameter, 4},
     {"_torch_cpp_jit_script_module_register_buffer", (DL_FUNC) &_torch_cpp_jit_script_module_register_buffer, 3},
+    {"_torch_cpp_jit_script_module_register_module", (DL_FUNC) &_torch_cpp_jit_script_module_register_module, 3},
     {"_torch_cpp_jit_script_module_to", (DL_FUNC) &_torch_cpp_jit_script_module_to, 3},
     {"_torch_cpp_jit_script_module_modules", (DL_FUNC) &_torch_cpp_jit_script_module_modules, 1},
     {"_torch_cpp_jit_script_module_children", (DL_FUNC) &_torch_cpp_jit_script_module_children, 1},
     {"_torch_cpp_jit_script_module_find_method", (DL_FUNC) &_torch_cpp_jit_script_module_find_method, 2},
     {"_torch_cpp_jit_script_method_call", (DL_FUNC) &_torch_cpp_jit_script_method_call, 2},
+    {"_torch_cpp_jit_script_module_new", (DL_FUNC) &_torch_cpp_jit_script_module_new, 2},
+    {"_torch_cpp_jit_script_module_add_constant", (DL_FUNC) &_torch_cpp_jit_script_module_add_constant, 3},
+    {"_torch_cpp_jit_script_module_add_method", (DL_FUNC) &_torch_cpp_jit_script_module_add_method, 2},
+    {"_torch_cpp_jit_script_module_find_constant", (DL_FUNC) &_torch_cpp_jit_script_module_find_constant, 2},
+    {"_torch_cpp_jit_script_module_save", (DL_FUNC) &_torch_cpp_jit_script_module_save, 2},
     {"_torch_test_stack", (DL_FUNC) &_torch_test_stack, 1},
     {"_torch_cpp_Tensor_storage", (DL_FUNC) &_torch_cpp_Tensor_storage, 1},
     {"_torch_cpp_Tensor_has_storage", (DL_FUNC) &_torch_cpp_Tensor_has_storage, 1},
@@ -35145,7 +35225,7 @@ static const R_CallMethodDef CallEntries[] = {
     {"_torch_cpp_torch_tensor_options", (DL_FUNC) &_torch_cpp_torch_tensor_options, 5},
     {"_torch_cpp_torch_tensor_options_print", (DL_FUNC) &_torch_cpp_torch_tensor_options_print, 1},
     {"_torch_test_fun_hello", (DL_FUNC) &_torch_test_fun_hello, 1},
-    {"_torch_cpp_trace_function", (DL_FUNC) &_torch_cpp_trace_function, 4},
+    {"_torch_cpp_trace_function", (DL_FUNC) &_torch_cpp_trace_function, 8},
     {"_torch_cpp_save_traced_fn", (DL_FUNC) &_torch_cpp_save_traced_fn, 2},
     {"_torch_cpp_jit_compilation_unit", (DL_FUNC) &_torch_cpp_jit_compilation_unit, 0},
     {"_torch_cpp_call_traced_fn", (DL_FUNC) &_torch_cpp_call_traced_fn, 2},