+++ NEWS.md
@@ -4,6 +4,7 @@
 - Added support for ScriptModule. Loaded JIT modules now operate as `nn_module`s. (#593)
 - Added a `jit_compile` function that allows compiling arbitrary TorchScript code into script function that can be serialized and executed. (#601)
 - Added Dockerfiles for CUDA 11.1 (#597)
+- Added `jit_trace` support for `nn_module` created from R. (#604)
 
 # torch 0.4.0